Feature/tutorial (#17)
* FIX : 게임창 리사이즈 수정

* ADD : 튜토리얼 모드 추가
